removed ManagementPositionCreditLine DAO

+++ src/net/sourceforge/fenixedu/applicationTier/Servico/credits/managementPosition/DeleteManagementPositionCreditLineService.java
@@ -1,24 +1,19 @@
 package net.sourceforge.fenixedu.applicationTier.Servico.credits.managementPosition;
 
-import net.sourceforge.fenixedu.applicationTier.Servico.framework.DeleteDomainObjectService;
-import net.sourceforge.fenixedu.domain.DomainObject;
+import net.sourceforge.fenixedu.applicationTier.Service;
+import net.sourceforge.fenixedu.applicationTier.Servico.exceptions.FenixServiceException;
 import net.sourceforge.fenixedu.domain.credits.ManagementPositionCreditLine;
 import net.sourceforge.fenixedu.persistenceTier.ExcepcaoPersistencia;
-import net.sourceforge.fenixedu.persistenceTier.IPersistentObject;
 
-public class DeleteManagementPositionCreditLineService extends DeleteDomainObjectService {
+public class DeleteManagementPositionCreditLineService extends Service {
 
-    protected Class getDomainObjectClass() {
-        return ManagementPositionCreditLine.class;
-    }
-
-    protected IPersistentObject getIPersistentObject() {
-        return persistentSupport.getIPersistentManagementPositionCreditLine();
-    }
-
-	protected void deleteDomainObject(DomainObject domainObject) throws ExcepcaoPersistencia {
-        ManagementPositionCreditLine managementPositionCreditLine = (ManagementPositionCreditLine) domainObject;
+    public void run(Integer managementPositionCreditLineID) throws ExcepcaoPersistencia, FenixServiceException {
+        ManagementPositionCreditLine managementPositionCreditLine = (ManagementPositionCreditLine) persistentObject.readByOID(
+                ManagementPositionCreditLine.class, managementPositionCreditLineID);
+        if (managementPositionCreditLine == null) {
+            throw new FenixServiceException("error.noManagementPositionCreditLine");
+        }
         managementPositionCreditLine.delete();
-	}
+    }
 
 }
